Move pull_repo to ef_conf_utils
