+++ tests/unit_tests/test_ef_utils.py
@@ -413,103 +413,6 @@ class TestEFUtils(unittest.TestCase):
       ef_utils.get_instance_aws_context(mock_ec2_client)
     self.assertIn("Did not find environment in role name:", exception.exception.message)
 
-  @patch('subprocess.check_call')
-  @patch('subprocess.check_output')
-  def test_pull_repo_ssh_credentials(self, mock_check_output, mock_check_call):
-    """
-    Tests pull_repo by mocking the subprocess.check_output to return git ssh credentials.
-
-    Args:
-      mock_check_output: MagicMock, returns valid git responses in order of being called, with the
-      repo coming from the ef_site_config.py
-
-    Returns:
-      None
-
-    Raises:
-      AssertionError if any of the assert checks fail
-    """
-    mock_check_output.side_effect = [
-      "user@" + EFConfig.EF_REPO.replace("/", ":", 1) + ".git",
-      EFConfig.EF_REPO_BRANCH
-    ]
-    try:
-      ef_utils.pull_repo()
-    except RuntimeError as exception:
-      self.fail("Exception occurred during test_pull_repo_ssh_credentials: " + exception.message)
-
-  @patch('subprocess.check_call')
-  @patch('subprocess.check_output')
-  def test_pull_repo_https_credentials(self, mock_check_output, mock_check_call):
-    """
-    Tests the pull_repo by mocking the subprocess.check_output to return git http credentials.
-
-    Args:
-      mock_check_output: MagicMock, returns valid git responses in order of being called, with the
-      repo coming from the ef_site_config.py
-
-    Returns:
-      None
-
-    Raises:
-      AssertionError if any of the assert checks fail
-    """
-    mock_check_output.side_effect = [
-      "origin\thttps://" + EFConfig.EF_REPO + ".git",
-      EFConfig.EF_REPO_BRANCH
-    ]
-    try:
-      ef_utils.pull_repo()
-    except RuntimeError as exception:
-      self.fail("Exception occurred during test_pull_repo_ssh_credentials: " + exception.message)
-
-  @patch('subprocess.check_output')
-  def test_pull_repo_incorrect_repo(self, mock_check_output):
-    """
-    Tests pull_repo to see if it throws an exception when the supplied repo doesn't match the one in
-    ef_site_config.py
-
-    Args:
-      mock_check_output: MagicMock, returns git responses with non matching repo names
-
-    Returns:
-      None
-
-    Raises:
-      AssertionError if any of the assert checks fail
-    """
-    mock_check_output.side_effect = [
-      "user@github.com:company/wrong_repo.git "
-      "other_user@github.com:company/wrong_repo.git"
-    ]
-    with self.assertRaises(RuntimeError) as exception:
-      ef_utils.pull_repo()
-    self.assertIn("Must be in", exception.exception.message)
-
-  @patch('subprocess.check_output')
-  def test_pull_repo_incorrect_branch(self, mock_check_output):
-    """
-    Tests pull_repo to see if it throws an error when the mocked check_output states it's on a branch
-    other than the one specified in ef_site_config.py
-
-    Args:
-      mock_check_output: MagicMock, returns some valid git responses, with the
-      repo coming from the ef_site_config.py, and then a non matching branch name
-
-    Returns:
-      None
-
-    Raises:
-      AssertionError if any of the assert checks fail
-    """
-    mock_check_output.side_effect = [
-      "user@" + EFConfig.EF_REPO.replace("/", ":", 1) + ".git",
-      "wrong_branch"
-    ]
-    with self.assertRaises(RuntimeError) as exception:
-      ef_utils.pull_repo()
-    self.assertIn("Must be on branch:", exception.exception.message)
-
   @patch('boto3.Session')
   def test_create_aws_clients(self, mock_session_constructor):
     """
